Move EventStreamVersion to the event_store module
